requirements.txt
Flask==2.3.3
cryptography==41.0.7
qrcode[pil]==7.4.2
Jinja2==3.1.2
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow>=9.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
requests>=2.28.0